backend/requirements.txt
Flask>=3.0
Flask-Cors>=4.0
Flask-JWT-Extended>=4.6
Flask-SQLAlchemy>=3.1
python-dotenv>=1.0
Werkzeug>=3.0
pip freeze > requirements.txt
